+++ src/components/views/kanban/index.ts
@@ -2,9 +2,6 @@
 export { KanbanList } from './components/KanbanList';
 export { KanbanListHeader } from './components/KanbanListHeader';
 export { KanbanCardsContainer } from './components/KanbanCardsContainer';
-export { AddCardButton } from './components/AddCardButton';
-export { AddListButton } from './components/AddListButton';
-export { KanbanDragOverlay } from './components/KanbanDragOverlay';
 
 // Hooks
 export { useKanbanDragAndDrop } from './hooks/useKanbanDragAndDrop';